pyproject.toml
[project]
name = "auto-story-tools"
version = "0.1.0"
description = "Autonomous story generation pipeline: seed → story bible → screenplay/novel"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Andrew Leveiss"},
]
keywords = ["ai", "story", "outline", "screenplay", "fiction", "writing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
    "mypy>=1.13",
]
screenplay = [
    "screenplain>=0.11.0",
]

[project.scripts]
auto-outline = "auto_outline.cli:main"
auto-screenplay = "auto_screenplay.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_outline", "src/auto_screenplay"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
